[refactor] step1 미션 피드백 반영
