+++ src/test/java/nextstep/subway/acceptance/LineSectionsAcceptanceTest.java
@@ -15,7 +15,7 @@ import static nextstep.subway.acceptance.StationSteps.지하철역_생성_요청
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철 구간 관리 기능")
-class LineSectionAcceptanceTest extends AcceptanceTest {
+class LineSectionsAcceptanceTest extends AcceptanceTest {
     private Long 신분당선;
 
     private Long 강남역;
